Close iostream before reconnecting
If too many pings are outstanding, `_send_ping` calls `_unbind`.  If
reconnecting is allowed, the client does reconnect, but the old
iostream is not being closed.  This results in 2 active connections to
the nats server.  After a while the 1st connection is kicked by the
server becuase it is stale.  Unfortunately, this late closing of the
first connection calls unbind again, so we get end up disconnecting
every few minutes, even though the current connection is fine.

The fix is to close the iostream before reconnecting.