Allow flusher loop to exit when disconnecting
If the `_flusher_loop` is blocked waiting on the `_flush_queue` when
we disconnect then the coroutine will just stay blocked there
forever.  This isn't critical, but cleaner to allow the flusher loop
to exit before spawning a new one - especially if we are reconnecting.

The `_read_loop` doesn't have this problem because it is blocked on
an iostream read which raises an exception when the iostream is closed.
This allows the read loop to exit.