Recover from flush timeouts on lost message
If the client sends a PING message and that message, or the replying
PONG gets lost somehow, this will cause all later `flush` attempts
to timeout even if the later PING-PONGs are fast enough.  This is
due to the `_pongs` list getting out of sync with the flush requests -
the oldest item in the list no longer refers to the current flush
attempt.

The fix is as follows:

- If the flush PING-PONG times out, set the future's result to False
  as an indication that this item in the `_pongs` list can be ignored
  later.
- When a PONG is received, remove all old items that have already timed
  out, and then one that hasn't yet timed out.

Note that `_pongs_received` counter doesn't refer to the number of PONGs
actually received, but rather the number of items removed from the
the `_pongs` list.